simpler sgdstore model
So far, I've only been able to get good results with an LSTM controller. I suspect that the LSTM is doing all the heavy lifting and not using the external net at all. I will have to look into that more.

After thousands of iterations with a batch size of 16, the cost got down to 0.42 (training) and 0.55 (validation).