bigger scale for one-hot inputs
Literally, the old input normalization scheme *prevented* learning. This new scheme scales the one-hot inputs by a factor of 16 and changes nothing else. Training works very fast now. In fact, initial learning is a lot faster than it was before (when the scale was only ~4x).

Why does this work? Well, it's not magic. By scaling the one-hot input, the network has an easier time "noticing" the change in one-hot input. Otherwise, the one-hot input is no more significant (initially) than a single pixel of the input image.